move the scene delegate in its own file
